added systemResult
