+++ packages/std/src/results/system.rs
@@ -1,0 +1,31 @@
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+use std::fmt;
+
+use crate::binary::Binary;
+use crate::errors::StdError;
+use crate::types::Empty;
+
+use super::attribute::Attribute;
+use super::cosmos_msg::CosmosMsg;
+
+pub struct SystemResponse<T = Empty>
+    where
+        T: Clone + fmt::Debug + PartialEq + JsonSchema,
+{
+    pub messages: Vec<CosmosMsg<T>>,
+    /// The attributes that will be emitted as part of a "wasm" event
+    pub attributes: Vec<Attribute>,
+    pub data: Option<Binary>,
+}
+impl<T> Default for SystemResponse {
+    fn default() -> Self {
+        SystemResponse {
+            messages: vec![],
+            attributes: vec![],
+            data: None,
+        }
+    }
+}
+
+pub type SystemResult<U = Empty> = Result<SystemResponse<U>, StdError>;
